Created separate class for HTML Report Test
