+++ src/test/java/com/redhat/exhort/integration/AnalysisTest.java
@@ -27,7 +27,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.IOException;
 import java.net.URI;
@@ -47,16 +46,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import com.gargoylesoftware.htmlunit.html.DomElement;
-import com.gargoylesoftware.htmlunit.html.DomNodeList;
-import com.gargoylesoftware.htmlunit.html.HtmlButton;
-import com.gargoylesoftware.htmlunit.html.HtmlHeading4;
-import com.gargoylesoftware.htmlunit.html.HtmlPage;
-import com.gargoylesoftware.htmlunit.html.HtmlTable;
-import com.gargoylesoftware.htmlunit.html.HtmlTableBody;
-import com.gargoylesoftware.htmlunit.html.HtmlTableDataCell;
-import com.gargoylesoftware.htmlunit.html.HtmlTableHeaderCell;
-import com.gargoylesoftware.htmlunit.html.HtmlTableRow;
 import com.redhat.exhort.api.PackageRef;
 import com.redhat.exhort.api.v4.AnalysisReport;
 import com.redhat.exhort.api.v4.DependencyReport;
@@ -426,121 +415,6 @@ public class AnalysisTest extends AbstractAnalysisTest {
     verifySnykRequest(OK_TOKEN);
   }
 
-  /**
-   * The generated HTML only has 1 vulnerability tab for Snyk. The quarkus-hibernate-orm has a
-   * private vulnerability that should be hidden and display the "Sign up" link to the user.
-   *
-   * <p>In order to expand the transitive table, it is required to click on the button contained in
-   * the <td>
-   */
-  @Test
-  public void testHtmlWithoutToken() {
-    stubAllProviders();
-
-    String body =
-        given()
-            .header(CONTENT_TYPE, CycloneDxMediaType.APPLICATION_CYCLONEDX_JSON)
-            .body(loadSBOMFile(CYCLONEDX))
-            .header("Accept", MediaType.TEXT_HTML)
-            .when()
-            .post("/api/v4/analysis")
-            .then()
-            .assertThat()
-            .statusCode(200)
-            .contentType(MediaType.TEXT_HTML)
-            .extract()
-            .body()
-            .asString();
-
-    var page = extractPage(body);
-    DomNodeList<DomElement> tables = page.getElementsByTagName("table");
-    assertEquals(1, tables.size());
-    var snykTable = tables.get(0);
-    var tbody = getTableBodyForDependency("io.quarkus:quarkus-hibernate-orm", snykTable);
-    assertNotNull(tbody);
-    page = expandTransitiveTableDataCell(tbody);
-    snykTable = page.getElementsByTagName("table").get(0);
-    tbody = getTableBodyForDependency("io.quarkus:quarkus-hibernate-orm", snykTable);
-
-    var issuesTable = getIssuesTable(tbody);
-    List<HtmlTableBody> tbodies = issuesTable.getByXPath(".//table//tbody");
-    var privateIssueTbody =
-        tbodies.stream()
-            .filter(
-                issuesTbody -> {
-                  List<HtmlTableDataCell> tds = issuesTbody.getByXPath("./tr/td");
-                  return tds.size() == 4;
-                })
-            .findFirst()
-            .get();
-    assertNotNull(privateIssueTbody);
-    HtmlTableDataCell td = privateIssueTbody.getFirstByXPath("./tr/td");
-    assertEquals(
-        "Sign up for a Snyk account to learn about the vulnerabilities found",
-        td.asNormalizedText());
-
-    verifySnykRequest(null);
-  }
-
-  /**
-   * This report contains both oss-index and snyk reports. So in order to show the Snyk report, we
-   * need to click on the tab. Then the quarkus-hibernate-orm having the unique vulnerability should
-   * appear without limitations.
-   *
-   * @throws IOException
-   */
-  @Test
-  public void testHtmlWithToken() throws IOException {
-    stubAllProviders();
-
-    var body =
-        given()
-            .header(CONTENT_TYPE, CycloneDxMediaType.APPLICATION_CYCLONEDX_JSON)
-            .body(loadSBOMFile(CYCLONEDX))
-            .header("Accept", MediaType.TEXT_HTML)
-            .header(Constants.SNYK_TOKEN_HEADER, OK_TOKEN)
-            .header(Constants.OSS_INDEX_USER_HEADER, OK_USER)
-            .header(Constants.OSS_INDEX_TOKEN_HEADER, OK_TOKEN)
-            .when()
-            .post("/api/v4/analysis")
-            .then()
-            .assertThat()
-            .statusCode(200)
-            .contentType(MediaType.TEXT_HTML)
-            .extract()
-            .body()
-            .asString();
-
-    var page = extractPage(body);
-    // Select the Snyk Source
-    HtmlButton snykSourceBtn = page.getFirstByXPath("//button[@aria-label='snyk source']");
-    assertNotNull(snykSourceBtn);
-    page = snykSourceBtn.click();
-
-    DomNodeList<DomElement> tables = page.getElementsByTagName("table");
-    assertEquals(2, tables.size());
-
-    var tbody = getTableBodyForDependency("io.quarkus:quarkus-hibernate-orm", tables.get(1));
-    assertNotNull(tbody);
-    page = expandTransitiveTableDataCell(tbody);
-    tables = page.getElementsByTagName("table");
-    tbody = getTableBodyForDependency("io.quarkus:quarkus-hibernate-orm", tables.get(1));
-
-    // TODO: figure out why the Snyk unique vulnerability is not being rendered in headless mode
-
-    // HtmlTable issuesTable = getIssuesTable(tbody);
-    // List<HtmlTableBody> tbodies = issuesTable.getByXPath(".//table//tbody");
-    // HtmlTableBody privateIssueTbody = tbodies.stream().filter(issuesTbody -> {
-    //   List<HtmlTableDataCell> tds = issuesTbody.getByXPath("./tr/td");
-    //   return tds.get(0).asNormalizedText().startsWith("SNYK");
-
-    // }).findFirst().get();
-    // assertNotNull(privateIssueTbody);
-
-    verifySnykRequest(OK_TOKEN);
-    verifyOssRequest(OK_USER, OK_TOKEN, false);
-  }
-
   @ParameterizedTest
   @ValueSource(strings = {"HTTP_1_1", "HTTP_2"})
   public void testMultipart_HttpVersions(String version) throws IOException, InterruptedException {
@@ -565,99 +439,6 @@ public class AnalysisTest extends AbstractAnalysisTest {
 
     verifySnykRequest(OK_TOKEN);
     verifyOssRequest(OK_USER, OK_TOKEN, false);
-  }
-
-  @Test
-  public void testHtmlUnauthorized() {
-    stubAllProviders();
-
-    var body =
-        given()
-            .header(CONTENT_TYPE, CycloneDxMediaType.APPLICATION_CYCLONEDX_JSON)
-            .body(loadSBOMFile(CYCLONEDX))
-            .header("Accept", MediaType.TEXT_HTML)
-            .header(Constants.SNYK_TOKEN_HEADER, INVALID_TOKEN)
-            .when()
-            .post("/api/v4/analysis")
-            .then()
-            .assertThat()
-            .statusCode(200)
-            .contentType(MediaType.TEXT_HTML)
-            .extract()
-            .body()
-            .asString();
-
-    var page = extractPage(body);
-    HtmlHeading4 heading = page.getFirstByXPath("//div[@class='pf-v5-c-alert pf-m-warning']/h4");
-    assertEquals(
-        "Warning alert:Snyk: Unauthorized: Verify the provided credentials are valid.",
-        heading.getTextContent());
-    assertTrue(page.getElementsByTagName("table").isEmpty());
-
-    verifySnykRequest(INVALID_TOKEN);
-    verifyNoInteractionsWithOSS();
-  }
-
-  @Test
-  public void testHtmlForbidden() {
-    stubAllProviders();
-
-    var body =
-        given()
-            .header(CONTENT_TYPE, CycloneDxMediaType.APPLICATION_CYCLONEDX_JSON)
-            .body(loadSBOMFile(CYCLONEDX))
-            .header("Accept", MediaType.TEXT_HTML)
-            .header(Constants.SNYK_TOKEN_HEADER, UNAUTH_TOKEN)
-            .when()
-            .post("/api/v4/analysis")
-            .then()
-            .assertThat()
-            .statusCode(200)
-            .contentType(MediaType.TEXT_HTML)
-            .extract()
-            .body()
-            .asString();
-
-    var page = extractPage(body);
-    HtmlHeading4 heading = page.getFirstByXPath("//div[@class='pf-v5-c-alert pf-m-warning']/h4");
-    assertEquals(
-        "Warning alert:Snyk: Forbidden: The provided credentials don't have the required"
-            + " permissions.",
-        heading.getTextContent());
-    assertTrue(page.getElementsByTagName("table").isEmpty());
-
-    verifySnykRequest(UNAUTH_TOKEN);
-    verifyNoInteractionsWithOSS();
-  }
-
-  @Test
-  public void testHtmlError() {
-    stubAllProviders();
-
-    var body =
-        given()
-            .header(CONTENT_TYPE, CycloneDxMediaType.APPLICATION_CYCLONEDX_JSON)
-            .body(loadSBOMFile(CYCLONEDX))
-            .header("Accept", MediaType.TEXT_HTML)
-            .header(Constants.SNYK_TOKEN_HEADER, ERROR_TOKEN)
-            .when()
-            .post("/api/v4/analysis")
-            .then()
-            .assertThat()
-            .statusCode(200)
-            .contentType(MediaType.TEXT_HTML)
-            .extract()
-            .body()
-            .asString();
-
-    var page = extractPage(body);
-    HtmlHeading4 heading = page.getFirstByXPath("//div[@class='pf-v5-c-alert pf-m-danger']/h4");
-    assertEquals(
-        "Danger alert:Snyk: Server Error: This is an example error", heading.getTextContent());
-    assertTrue(page.getElementsByTagName("table").isEmpty());
-
-    verifySnykRequest(ERROR_TOKEN);
-    verifyNoInteractionsWithOSS();
   }
 
   @Test
@@ -731,51 +512,5 @@ public class AnalysisTest extends AbstractAnalysisTest {
             .orElse(null);
     assertNotNull(dep);
     return dep;
-  }
-
-  private HtmlTableBody getTableBodyForDependency(String depRef, DomElement table) {
-    List<HtmlTableBody> tbodies = table.getByXPath(".//tbody");
-    return tbodies.stream()
-        .filter(
-            tbody -> {
-              HtmlTableHeaderCell th = tbody.getFirstByXPath("./tr/th");
-              return th.asNormalizedText().equals(depRef);
-            })
-        .findFirst()
-        .orElse(null);
-  }
-
-  private HtmlPage expandTransitiveTableDataCell(HtmlTableBody tbody) {
-    return expandTableDataCell(tbody, "Transitive Vulnerabilities");
-  }
-
-  private HtmlPage expandDirectTableDataCell(HtmlTableBody tbody) {
-    return expandTableDataCell(tbody, "Direct Vulnerabilities");
-  }
-
-  private HtmlPage expandTableDataCell(HtmlTableBody tbody, String dataLabel) {
-    HtmlTableDataCell td =
-        tbody.getFirstByXPath(String.format("./tr/td[@data-label='%s']", dataLabel));
-    if (td.getAttribute("class").contains("pf-m-expanded")) {
-      return tbody.getHtmlPageOrNull();
-    }
-    try {
-      HtmlButton button = td.getFirstByXPath("./button");
-      return button.click();
-    } catch (IOException e) {
-      fail("Unable to click on td for data-label: " + dataLabel, e);
-      return null;
-    }
-  }
-
-  private HtmlTable getIssuesTable(HtmlTableBody dependencyTable) {
-    List<HtmlTableRow> rows = dependencyTable.getByXPath("./tr");
-    if (rows.size() != 2) {
-      fail(
-          "Expected table to have 2 <tr>. One for the dependency name and another for the"
-              + " vulnerabilities. Found: "
-              + rows.size());
-    }
-    return rows.get(1).getFirstByXPath("//table");
   }
 }
